refactor: functions into dedicated files
