+++ packages/shared/lib/core/wallet/actions/index.ts
@@ -1,8 +1,13 @@
 export * from './claimActivity'
-export * from './sendOutput'
+export * from './getAccountAssetsForSelectedAccount'
 export * from './hideActivity'
 export * from './linkActivityAndClaimingTransaction'
-export * from './loadAccountActivies'
+export * from './loadAccountActivities'
 export * from './loadAccountActivitiesFromOutputs'
 export * from './loadAccountActivitiesFromTransactions'
+export * from './refreshAccountAssetsForActiveProfile'
+export * from './requestAndStorePersistedAsset'
+export * from './sendOutput'
+export * from './setAsyncActivitiesToClaimed'
 export * from './setAsyncStatusOfAccountActivities'
+export * from './tryGetAndStoreAssetFromPersistedAssets'
