+++ packages/shared/lib/core/wallet/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './buildPersistedAssetFromIrc30Metadata'
