+++ packages/shared/lib/core/wallet/utils/index.ts
@@ -1,16 +1,21 @@
+// Files
 export * from './convertEd25519ToBech32'
 export * from './formatTokenAmountBestMatch'
 export * from './formatTokenAmountDefault'
 export * from './formatTokenAmountPrecise'
-export * from './getSubjectFromAddress'
+export * from './generateRawAmount'
 export * from './getActivityType'
+export * from './getAssetFromPersistedAssets'
+export * from './getIrc30MetadataFromFoundryOutput'
+export * from './getMetadataFromFoundryOutput'
 export * from './getOutputOptions'
 export * from './getSenderAddressFromUnlockCondition'
-export * from './getTokenMetadataFromFoundryOutput'
-export * from './generateRawAmount'
+export * from './getSubjectFromAddress'
 export * from './isSubjectInternal'
 export * from './parseRawAmount'
+export * from './validateIrc30Metadata'
 
+// Folders
 export * from './outputs'
 export * from './send'
 export * from './transactions'
